small update

+++ packages/project/src/generators/project/impl/finalization.ts
@@ -1,5 +1,7 @@
 import { execCommand } from '@gmnx/internal-util';
 
-export async function finalization(): Promise<void> {
-  await execCommand('npm install -f');
+export async function finalization(dryRun: boolean): Promise<void> {
+  if (!dryRun) {
+    await execCommand('npm install -f');
+  }
 }
